Split fixtures and tests to different files
