+++ tests/test_currency.py
@@ -1,46 +1,8 @@
-import pytest
-import requests
 from app.internal.currency import is_valid_currency_api_response
 from fastapi import status
 
 
 CURRENCY = '/currency'
-
-
-@pytest.fixture
-def requests_mock_good(monkeypatch):
-    class MockResponse:
-        def __init__(self):
-            self.status_code = 200
-
-    def mock_requests_get_good(url):
-        return MockResponse()
-
-    monkeypatch.setattr(requests, 'get', mock_requests_get_good)
-
-
-@pytest.fixture
-def requests_mock_bad(monkeypatch):
-    class MockResponse:
-        def __init__(self):
-            self.status_code = 403
-
-    def mock_requests_get_bad(url):
-        return MockResponse()
-
-    monkeypatch.setattr(requests, 'get', mock_requests_get_bad)
-
-
-@pytest.fixture
-def requests_mock_error(monkeypatch):
-    class MockResponse:
-        def __init__(self):
-            raise requests.ConnectionError("Test")
-
-    def mock_requests_error(url):
-        return MockResponse()
-
-    monkeypatch.setattr(requests, 'get', mock_requests_error)
 
 
 def test_reponse_bad(requests_mock_bad):
